Mark our explicit fall through so -Wextra will work in gcc 7
gcc 7 introduces detection of fall-through behavior in switch/case
statements, and will warn if -Wimplicit-fallthrough is present and there
is no comment stating that the fall-through is intentional.  This is
also triggered by -Wextra, as it enables -Wimplicit-fallthrough=1.

This patch adds the comment in the one place we use fall-through.

Signed-off-by: Peter Jones <pjones@redhat.com>